requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.4
cryptography==3.2.1
decorator==4.4.2
mysqlclient==2.0.1
pycparser==2.20
python-dotenv==0.15.0
python-telegram-bot==13.1
pytz==2020.4
six==1.15.0
SQLAlchemy==1.3.20
tornado==6.1
tzlocal==2.1
